requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
